refactor: use `casdoorsdk` as the pkg name
