+++ casdoorsdk/sms.go
@@ -12,25 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package auth
+package casdoorsdk
 
 import (
 	"encoding/json"
 	"fmt"
 )
 
-type emailForm struct {
-	Title     string   `json:"title"`
+type smsForm struct {
 	Content   string   `json:"content"`
-	Sender    string   `json:"sender"`
 	Receivers []string `json:"receivers"`
 }
 
-func SendEmail(title string, content string, sender string, receivers ...string) error {
-	form := emailForm{
-		Title:     title,
+func SendSms(content string, receivers ...string) error {
+	form := smsForm{
 		Content:   content,
-		Sender:    sender,
 		Receivers: receivers,
 	}
 	postBytes, err := json.Marshal(form)
@@ -38,7 +34,7 @@ func SendEmail(title string, content string, sender string, receivers ...string)
 		return err
 	}
 
-	resp, err := doPost("send-email", nil, postBytes, false)
+	resp, err := doPost("send-sms", nil, postBytes, false)
 	if err != nil {
 		return err
 	}
